Add background view for empty list; disable rotation and dark mode

+++ UserManager/UserManager/Modules/UserList/UserListViewController.swift
@@ -86,16 +86,43 @@ private extension UserListViewController {
             .sink { [unowned self] event in
                 switch event {
                     case .setUsers(let users):
-                        var snapshot = dataSource.snapshot()
-                        snapshot.deleteAllItems()
-                        snapshot.appendSections([.users])
-                        snapshot.appendItems(users, toSection: .users)
-                        dataSource.apply(snapshot)
+                        if users.isEmpty {
+                            self.contentTableView.setEmptyMessage(AppConstants.UserList.Message.emptyList.rawValue)
+                        } else {
+                            self.contentTableView.restore()
+                            var snapshot = dataSource.snapshot()
+                            snapshot.deleteAllItems()
+                            snapshot.appendSections([.users])
+                            snapshot.appendItems(users, toSection: .users)
+                            dataSource.apply(snapshot)
+                        }
                 }
             }.store(in: &cancellables)
     }
     
     @objc private func onAdddUserTapped() {
         router.routeTo(.addUser)
+    }
+}
+
+private extension UITableView {
+    
+    var emptyMessageLabel: UILabel {
+        let label = UILabel(frame: CGRect(x: 0, y: 0, width: bounds.size.width, height: bounds.size.height))
+        label.textColor = .darkText
+        label.numberOfLines = 0
+        label.textAlignment = .center
+        label.sizeToFit()
+        return label
+    }
+    
+    func setEmptyMessage(_ message: String) {
+        let label = emptyMessageLabel
+        label.text = message
+        backgroundView = label
+    }
+
+    func restore() {
+        backgroundView = nil
     }
 }
